Improved constant folding
The interpreter is now capable of shrinking, if not removing the
use of variable declarations. Consider the following example:

> let x = 2; x

In the example, we can see that the value of x can be substituted into
the body following the let. Thus, it simplifies to:

> let x = 2; 2

In this commit, I have added the ability to further simplify this by
checking for variable usage in a program. For instance, because x is
not used to compute 2, we can remove the declaration, and with it, the
entire let expressiin:

> 2

Hence, the expression requires far less overhead to compute. In this
example, the benefit is marginal. However, when used in loops, the
effects could be massive.