Added explicit thunking
A thunk is a principle of lazy evaluation in which the evaluation of
an expression can be saved until it is required. In some languages,
this can be done explicitly. However, this language uses explicit
declarations of thunk expressions.

For instance, consider the expression "let f(x) = f(x); f(2); true".
This will never halt due to the fact that any call to f will infinitely
recurse. This is inconvenient because neither the value nor the side effects
of f(2) are used in the computation of any results. So, we can wrap the
function call in a thunk, and evaluate "let f(x) = f(x); thunk {f(2)}; true".
Note that using a thunk on an expression with side effects will have the
effect of preventing those side effects from occuring. Hence, the thunk
expressions should be used with great care.