Implemented basic modules
A file can be evaluated and imported via the 'import' keyword.
This will load a .lom file matching the given name, and then evaluate
it as its own program. The outcome of the program will be stored in
a variable based on the usage. For instance, 'import module' creates a
variable module based on running 'module.lom', and 'import module as mod'
will create run 'module.lom' and store the result in a variable called
mod. This syntax is admittedly very similar to the Python syntax.
However, it works, and hence I'll keep it for the moment.