Added dictionaries
In Lomda, dictionaries are defined by providing identifier fields
paired with evaluable expressions. The syntax is very much alike
the Javascript syntax for defining objects, due to a lack of
imagination on my part. I don't know if I'll change it, but if
I think of something, I'll be open to it.