Improved notion of stdlib using implementable exp
There is now an expression that can be defined using a collection
of functions. This allows for functions with no memory that simply
utilize a prewritten program to perform computation. This should
also make it easier to develop standard libraries, as I can simply
write a body function and pass it to the function.