Long desired tweak to matrix multiplication
I have altered matrix multiplication to loosen part of the restrictions
on multiplication. This is because in differentiation, the product rule
will not be able to evaluate if values beyond matrices can't be operated
on. Hence, I loosened this from requiring both be matrix or lower to at
least one of them being matrix or lower.