Differentiation changes
I have implemented an adjustment to the chain rule defined for the
ApplyExp. Now, the expression will perform a reduction to a point
where a dot product can be used. This was done in response to another
feature in this commit that adds differentiation of the trace and
transpose functions from the linalg library.